% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{calc_window_binary}
\alias{calc_window_binary}
\title{Calculate Sliding Window Binary}
\usage{
calc_window_binary(inwindow)
}
\arguments{
\item{inwindow}{A window, sub-data.table of the main data.table}
}
\value{
+1 if acceleration, -1 if deacceleration
}
\description{
Calculates the sliding window smoothing for a window, returning binary +1/-1
depending on if the acceleration has increased in the window or decreased.
Calculated as increase if the mean of the window is higher than the first
element and vice versa for deacceleration.
}
