% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{calc_window_mean}
\alias{calc_window_mean}
\title{Calculate Sliding Window Mean}
\usage{
calc_window_mean(inwindow, threshold)
}
\arguments{
\item{inwindow}{A window, sub-data.table of the main data.table}

\item{threshold}{A threshold}
}
\value{
Returns mean acceleration of window if the difference in acceleration
between the first element and the mean of the window is greater than or equal
to threshold, otherwise returns the first element
}
\description{
Calculates the sliding window smoothing for a window, based on the mean
acceleration of the window and the threshold.
}
