% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_dt}
\alias{load_dt}
\title{Load Data Table}
\usage{
load_dt(infile)
}
\arguments{
\item{infile}{Path to data to read in}
}
\value{
Returns data in data.table format
}
\description{
Loads the data to be used in the sliding window, does nothing if file
does not exist.
}
