% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sw_plot}
\alias{sw_plot}
\title{Sliding Window Plot}
\usage{
sw_plot(indata, plot_mean, plot_binary, plot_bg)
}
\arguments{
\item{indata}{Takes data in the form of a data.table}

\item{plot_mean}{Plots the mean line if the algorithm was used}

\item{plot_binary}{Plots the binary line if the algorithm was used}

\item{plot_bg}{Plots the background euclidean line if used}
}
\value{
Returns the plot made
}
\description{
Plots the data after algorithms have been applied to smoothen the data.
}
